Added check for templates.tar.gz

+++ DemoUpdateScripts/03-updateBackupTemplatesFromWeb.sh
@@ -11,20 +11,23 @@ cd ${NUXEO_DEMO_PARENT_DIR}/${NUXEO_DEMO_DIR} || exit 1
 
 
 # check if templates file exists on demo site server
-wget -S --spider  "${NUXEO_DEMO_SITE}${NUXEO_DEMO_TEMPLATES}.tar.gz"|| exit 1 
+wget -S --spider  "${NUXEO_DEMO_SITE}${NUXEO_DEMO_TEMPLATES}.tar.gz"
 
-wget -N "${NUXEO_DEMO_SITE}${NUXEO_DEMO_TEMPLATES}.tar.gz"|| exit 1
-
-if test -d ${NUXEO_DEMO_PARENT_DIR}/${NUXEO_DEMO_DIR}/${NUXEO_DEMO_TEMPLATES}
+if [ $? -eq 0 ]
 then
-	echo "Older templates found, resetting it"
-	rm -r ${NUXEO_DEMO_PARENT_DIR}/${NUXEO_DEMO_DIR}/${NUXEO_DEMO_TEMPLATES}
-fi
 
-echo "Extracting templates"
-tar xzf ${NUXEO_DEMO_TEMPLATES}.tar.gz || exit 1
-wait
-echo "Templates extracted"
-
-
-
+  wget -N "${NUXEO_DEMO_SITE}${NUXEO_DEMO_TEMPLATES}.tar.gz"|| exit 1
+  
+  if test -d ${NUXEO_DEMO_PARENT_DIR}/${NUXEO_DEMO_DIR}/${NUXEO_DEMO_TEMPLATES}
+  then
+  	echo "Older templates found, resetting it"
+  	rm -r ${NUXEO_DEMO_PARENT_DIR}/${NUXEO_DEMO_DIR}/${NUXEO_DEMO_TEMPLATES}
+  fi
+  
+  echo "Extracting templates"
+  tar xzf ${NUXEO_DEMO_TEMPLATES}.tar.gz || exit 1
+  wait
+  echo "Templates extracted"
+  
+fi 
+
